requirements.txt
Django==1.11.5
djangorestframework==3.7.7
django-cors-headers==2.1.0
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.3
